Added PurchasedItem model
This allows us to keep track of what a user buys in which quantity. This is
useful because in most cases we need to make assumptions like "If the user
bought a subscription less than a year ago, allow access to the site". We
can now easily do that without needing to update the UserProfile model all
the time - and we could show a list of transactions (generate invoices) to
our users.